Renamed Database namespaces and classes
